Версия с непонятным циклом

+++ DropItViewController.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-class DropItViewController: UIViewController
+class DropItViewController: UIViewController, UIDynamicAnimatorDelegate
 {
 
     @IBOutlet weak var gameView: UIView!
@@ -38,13 +38,23 @@ class DropItViewController: UIViewController
     
     
 //    Количество фигурок в ряд
-    var dropsPerRow = 10
+    var dropsPerRow = 18
+    var dropsPerColumn = 10
+    
+//    Должен делить высоту фрейма на дропсайз (ширина она же высота кубика) и получать значение переменной
     
 //    Высчитываем размер объектов
     var dropSize: CGSize{
         let size = gameView.bounds.size.width / CGFloat(dropsPerRow)
-        return CGSize(width: size, height: size)
+        return CGSize(width: size, height: size)  /* Ширина и высота кубика */
     }
+    
+    var dropSizeHeight: CGSize{
+        let sizeHeight = gameView.bounds.size.height / CGFloat(dropsPerColumn) /* Делю высоту фрейма на количество кубиков (должен не на количество кубиков, а на относительную их высоту равную полученной ранее ширине) */
+        let sizeWidth = gameView.bounds.size.width / CGFloat(dropsPerRow)
+        return CGSize(width: sizeWidth, height: sizeHeight)
+    }
+    
     
     @IBAction func btn(sender: UIButton) {
     
@@ -57,12 +67,12 @@ class DropItViewController: UIViewController
         
     }
 
-    
+    var indexOf: Int = 0
 
 //    Функция отрисовки фигур и применения анимации
     func drop(){
      
-           var frame = CGRect(origin: CGPointZero, size: dropSize)
+        var frame = CGRect(origin: CGPointZero, size: dropSize)
         frame.origin.x = CGFloat.random(dropsPerRow) * dropSize.width
         
         let dropView = UIView(frame: frame)
@@ -73,28 +83,31 @@ class DropItViewController: UIViewController
         gravity.addItem(dropView)
         collider.addItem(dropView) 
         
+
         
- 
-        
-//        Цикл
-        for count in 1...2 {
-            
-            let delay = Int64( Double(count) * Double(NSEC_PER_SEC) * 5)
-            let dispatchTime = dispatch_time(DISPATCH_TIME_NOW, delay)
-            dispatch_after(dispatchTime, dispatch_get_main_queue()) {
-                self.drop()
-                
-                
+//      Цикл генерации фигур
+    
+        for cnt in 1...1
+        {
+            if indexOf < 10
+            {
+                for var indexOf = 0; indexOf == 10; ++indexOf
+                {
+                    let delay = Int64( Double(cnt)/100 * Double(NSEC_PER_SEC) )
+                    let dispatchTime = dispatch_time(DISPATCH_TIME_NOW, delay)
+                    dispatch_after(dispatchTime, dispatch_get_main_queue())
+                        {
+                            self.drop()
+                    }
+                }
+            }
+            else
+            {
+                break
             }
         }
         
     }
-    
-    
-    
-    
-    
-    
 }
 
 
@@ -110,13 +123,13 @@ private extension CGFloat{
 // Создаем метод для рандомной раскраски объектов
 private extension UIColor{
     class var random: UIColor {
-        switch arc4random()%5 {
-        case 0: return UIColor.greenColor()
-        case 1: return UIColor.blueColor()
-        case 2: return UIColor.orangeColor()
-        case 3: return UIColor.redColor()
-        case 4: return UIColor.magentaColor()
-        default: return UIColor.blackColor()
+        switch arc4random()%4 {
+        case 0: return UIColor.blackColor()
+        case 1: return UIColor.lightGrayColor()
+        case 2: return UIColor.grayColor()
+        case 3: return UIColor.darkGrayColor()
+//        case 4: return UIColor.magentaColor()
+        default: return UIColor.purpleColor()
         }
     }
 }
